Create Model prediction
The file imported for training needs to be in xlsx format, consistent with the training, with the molecule name (or codename) in the first column, the 1024-bit ECFP fingerprint features in the second to penultimate column, and the annotation labels in the last column. The format of the prediction file also needs to be exactly the same, except for the label in the last column. this means converting the file to be predicted to the same 1024-bit ECFP molecular fingerprint feature.